Telegram Trade Copier bot/requirements.txt
python-telegram-bot ==13.15
requests==2.26.0
python-dotenv==0.19.2